% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_CONFUSION_MATRIX.R
\name{cutoff.palette}
\alias{cutoff.palette}
\title{Palette of cutoff values that minimize and maximize metrices from the confusion matrix}
\usage{
cutoff.palette(predictions, observed, min.pct.obs = 0.05, min.pct.def = 0.01)
}
\arguments{
\item{predictions}{Model predictions.}

\item{observed}{Observed values of target variable.}

\item{min.pct.obs}{Minimum percentage of observations. Used to select boundaries of cutoff values. Default value is 0.05.}

\item{min.pct.def}{Minimum percentage of default. Used to select boundaries of cutoff values. Default value is 0.01.}
}
\value{
The command \code{cutoff.palette} returns data frame with minimum and maximum values of each confusion
matrix metric along with optimized cutoff itself.
}
\description{
\code{cutoff.palette} returns confusion matrix along with accompanied performance metrics.
}
