% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_ENSEMBLE_BLOCKS.R
\name{ensemble.blocks}
\alias{ensemble.blocks}
\title{Ensemble blocks regression}
\usage{
ensemble.blocks(
  method,
  target,
  db,
  coding = "WoE",
  blocks,
  p.value = 0.05,
  miv.threshold = 0.02,
  m.ch.p.val = 0.05
)
}
\arguments{
\item{method}{Regression method applied on each block.
Available methods: \code{"stepMIV"}, \code{"stepFWD"}, \code{"stepRPC"}, \code{"stepFWDr"}, and \code{"stepRPCr"}.}

\item{target}{Name of target variable within \code{db} argument.}

\item{db}{Modeling data with risk factors and target variable.}

\item{coding}{Type of risk factor coding within the model. Available options are: \code{"WoE"} and
\code{"dummy"}. If \code{"WoE"} is selected, then modalities of the risk factors are replaced
by WoE values, while for \code{"dummy"} option dummies (0/1) will be created for \code{n-1}
modalities where \code{n} is total number of modalities of analyzed risk factor.}

\item{blocks}{Data frame with defined risk factor groups. It has to contain the following columns: \code{rf} and
\code{block}.}

\item{p.value}{Significance level of p-value for the estimated coefficient. For \code{WoE} coding this value is
is directly compared to p-value of the estimated coefficient, while for \code{dummy} coding
multiple Wald test is employed and its p-value is used for comparison with selected threshold (\code{p.value}).
This argument is applicable only for \code{"stepFWD"} and \code{"stepRPC"} selected methods.}

\item{miv.threshold}{MIV (marginal information value) entrance threshold applicable only for code{"stepMIV"} method.
Only the risk factors with MIV higher than the threshold are candidate for the new model.
Additional criteria is that MIV value should significantly separate
good from bad cases measured by marginal chi-square test.}

\item{m.ch.p.val}{Significance level of p-value for marginal chi-square test applicable only for code{"stepMIV"} method.
This test additionally supports MIV value of candidate risk factor for final decision.}
}
\value{
The command \code{embeded.blocks} returns a list of three objects.\cr
The first object (\code{model}) is the list of the models of each block (an object of class inheriting from \code{"glm"}).\cr
The second object (\code{steps}), is the data frame with risk factors selected from the each block.\cr
The third object (\code{dev.db}), returns the list of block's model development databases.\cr
}
\description{
\code{ensemble.blocks} performs blockwise regression where the predictions of each blocks' model are
integrated into a final model. The final model is estimated in the form of logistic regression without
any check of the estimated coefficients (e.g. statistical significance or sign of the estimated coefficients).
}
\examples{
suppressMessages(library(PDtoolkit))
data(loans)
#create risk factor priority groups
rf.all <- names(loans)[-1]
set.seed(22)
blocks <- data.frame(rf = rf.all, block = sample(1:3, length(rf.all), rep = TRUE))
blocks <- blocks[order(blocks$block), ]
blocks
#method: stepRPCr
res <- ensemble.blocks(method = "stepRPCr", 
		      target = "Creditability",
		      db = loans,
		      blocks = blocks, 
		      p.value = 0.05)
names(res)
nb <- length(res[["models"]])
res$models[[nb]]
auc.model(predictions = predict(res$models[[nb]], type = "response", 
				    newdata = res$dev.db[[nb]]),
      observed = res$dev.db[[nb]]$Creditability)
}
\references{
Anderson, R.A. (2021). Credit Intelligence & Modelling, Many Paths through the Forest of Credit Rating and Scoring,
OUP Oxford
}
\seealso{
\code{\link{staged.blocks}}, \code{\link{embedded.blocks}}, \code{\link{stepMIV}}, \code{\link{stepFWD}},
\code{\link{stepRPC}}, \code{\link{stepFWDr}} and \code{\link{stepRPCr}}.
}
