% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loans}
\alias{loans}
\title{German Credit Data}
\format{
An object of class \code{data.frame} with 1000 rows and 21 columns.
}
\source{
\url{https://online.stat.psu.edu/stat857/node/215/}
}
\usage{
loans
}
\description{
The German Credit Data contains data on 20 variables and the classification whether
an applicant is considered a Good or a Bad credit risk for 1000 loan applicants.
Name of the columns are used as give in the source file.
Note that subset of those data is available also in 'monobin' package (gcd) and used for some examples in 'PDtoolkit' package.
}
\keyword{datasets}
