% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_BIVARIATE_ANALYSIS.R
\name{replace.woe}
\alias{replace.woe}
\title{Replace modalities of risk factor with weights of evidence (WoE) value}
\usage{
replace.woe(db, target)
}
\arguments{
\item{db}{Data frame of categorical risk factors and target variable supplied for WoE coding.}

\item{target}{Name of target variable within \code{db} argument..}
}
\value{
The command \code{replace.woe} returns the list of two data frames. The first one contains WoE replacement
of analyzed risk factors' modalities, while the second data frame reports results of above
validations regarding class of the risk factors and number of modalities.
}
\description{
\code{replace.woe} replaces modalities of risk factor with calculated WoE value. This function process only
categorical risk factors, thus it is assumed that numerical risk factors are previously categorized.
Additional info report (second element of function output - \code{info} data frame), if produced, includes:
\itemize{
\item rf: Risk factor name.
\item reason.code: Reason code takes value 1 if inappropriate class of risk factor is identified, while
for check of maximum number of categories it takes value 2.
\item comment: Reason description.
}
}
\examples{
suppressMessages(library(monobin))
data(gcd)
#categorize numeric risk factor
gcd$maturity.bin <- ndr.bin(x = gcd$maturity, y = gcd$qual, y.type = "bina")[[2]]
gcd$amount.bin <- ndr.bin(x = gcd$amount, y = gcd$qual, y.type = "bina")[[2]]
gcd$age.bin <- ndr.bin(x = gcd$age, y = gcd$qual, y.type = "bina")[[2]]
head(gcd)
#replace modalities with WoE values
woe.rep <- replace.woe(db = gcd, target = "qual")
#results overview
head(woe.rep[[1]])
woe.rep[[2]]
}
